Create missing Lutris and Steam directories
Some directories are not present by default. For example, Steam requires
the user to manually create the `compatibilitytools.d` folder or Lutris
does not create a runner config file for a Wine setup if the settings
are never changed.

Therefore, GE-Man now ensures that the relevant directories are present
to prevent errors due to missing directories in a path.