Do not treat adding an existing version as an error
Rust returns a non-zero exist code when returning an error result.

Currently, when adding an existing version an error is returned with
anyhow. This makes it hard to check for a successful `add` operation in
scripting by checking the success code.

A possible use case might be to download the latest version every day by
running `geman add -p --apply`.