Remove pointless AppConfigPaths struct
The `list` and `remove` commands in the ui module relied on the
AppConfigPaths struct that contained the config file path for Steam and
Lutris. This struct would then be passed into the relevant remove and
list methods. However, this approach made no sense, because the
TerminalWriter already has a dependency on a PathConfiguration trait,
therefore, the dependency present in the TerminalWriter could just be
used to perform the same task of getting the required config file paths.

Furthermore, the AppConfigPaths struct used a struct implementing the
PathConfiguration trait for fetching the config file paths. Making the
entire reason for this struct even more pointless.