Reduce ApplyCommandInput content & rework GivenVersion to use a generic
ApplyCommandInput now stores a managed version which simplifies future
interactions and tests in the command_execution module.

GivenVersion now uses a generic instead of a Box. A generic makes more
sense for the intended purpose. Once a GivenVersion holds a specific
value of a type it is not expected to change.