Code rework for label feature
This commits add the label attribute to the structs ManagedVersion and
Version. Additionally, it refactors many places in the code to make it
compile again, however it does not guarantee that the code is logically
valid and that tests work.

Furthermore, it improves the test setup by extracting fixtures into a
`fixture` module in `lib.rs` so that when future changes occur the
fixtures only need to be changed at a single location and not across the
entire code base in different modules.